requirements.txt
apache-airflow-providers-sqlite==2.1.3
pytest==7.1.2
avro==1.11.0
black==22.3.0
isort==5.10.1
autoflake==1.4
pandas == 2.0.3